requirements.txt
asgiref==3.6.0
Django==4.1.4
django-webpack-loader==1.8.1
psycopg2-binary==2.9.5
python-dotenv==0.21.1
sqlparse==0.4.3
gunicorn==20.1.
dj-database-url>=1.0,<2.0
whitenoise>=6.0,<7.0
djangorestframework==3.14.0
drf-yasg==1.21.5